# Conflicts:
#	include/pika_admin.h
#	include/pika_bit.h
#	include/pika_command.h
#	include/pika_geo.h
#	include/pika_hash.h
#	include/pika_kv.h
#	include/pika_list.h
#	include/pika_pubsub.h
#	include/pika_set.h
#	include/pika_transaction.h
#	include/pika_zset.h
#	src/pika_command.cc

+++ include/pika_admin.h
@@ -388,7 +388,7 @@ class TimeCmd : public Cmd {
 
 class LastsaveCmd : public Cmd {
  public:
-  LastsaveCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  LastsaveCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override {};
   void Merge() override {};
